+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -7,6 +7,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import kitchenpos.product.domain.Product;
 
 @Entity
 public class MenuProduct {
