+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,9 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.application.response.ProductResponse;
-import kitchenpos.domain.Product;
+import kitchenpos.product.response.ProductResponse;
+import kitchenpos.product.domain.Product;
 import kitchenpos.domain.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
