+++ src/test/java/kitchenpos/integration/ProductIntegrationTest.java
@@ -6,8 +6,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import kitchenpos.integration.api.ProductApi;
 import kitchenpos.integration.utils.MockMvcResponse;
-import kitchenpos.ui.request.ProductCreateRequest;
-import kitchenpos.application.response.ProductResponse;
+import kitchenpos.product.ui.request.ProductCreateRequest;
+import kitchenpos.product.response.ProductResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
